requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-community>=0.0.10
streamlit>=1.24.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.2
langsmith>=0.0.52
faiss-cpu>=1.7.4
pydantic>=2.0.0
loguru>=0.7.0
pytest>=7.4.0
python-jose>=3.3.0
requests>=2.31.0
typing-extensions>=4.5.0
google-api-python-client>=2.0.0
openai>=1.0.0 


